Creating base class for nonlocal material averaged vector postprocessor at crack tip
