+++ modules/solid_mechanics/test/tests/crack_front_stress/gold/crack_front_nonlocal_stress_out_CrackFrontNonlocalStress_0001.csv
@@ -1,4 +1,4 @@
-crack_tip_stress,id,x,y,z
+crack_tip_generic_stress,id,x,y,z
 10.012781609899,0,-1.2246467991474e-17,-2.4492935982947e-17,0.2
 10,0.041811385307061,-1.0677810676341e-17,-0.041582338163552,0.19562952014676
 10,0.083622770614123,-1.2055066943765e-17,-0.08134732861516,0.18270909152852
